requirements.txt
Django==1.10
psycopg2==2.5.4
sqlparse==0.1.13
static3==0.5.1
dj-database-url==0.4.1

djangorestframework==3.4.6
drfdocs==0.0.11
-e git+https://github.com/decko/django-rest-localflavor#egg=django-localflavor
django-cors-headers==1.2.2
django-filter==0.14.0
geopy==1.11.0
drf-oidc-auth==0.8

pendulum>=0.5.4
model-mommy==1.2.6

# Pacotes nescessarios para a suite de testes
pytest==2.9.2
pytest-django==2.9.1
pytest-spec==1.0.1
pytest-sugar==0.7.1
pytest-xdist==1.14
pytest-cov==2.3.0
